[feat] 서버통신 mutate작성. 기능 확인. #28
